+++ ExpressionEval/Nodes/DivideNode.cs
@@ -3,6 +3,6 @@
 		public DivideNode() {}
 		public DivideNode(Node left, Node right) : base(left, right) {}
 
-		public override int Value => Left / Right;
+		public override float Value => Left / Right;
 	}
 }
